refactor: architecture polish — eliminate globals, extract background services, split files
- Remove dead pubsub.GetGlobalPubSub/SetGlobalPubSub (defined but never called)
- Inject rate limiter into functions.Handler via SetRateLimiter() instead
  of global ratelimit.GetGlobalStore()
- Add RateLimiter field to ServiceRegistry (same pattern as PubSub)
- Extract BackgroundServicesModule (starts realtime listener, schedulers,
  jobs manager, RPC scheduler via leader election)
- Extract leaderElect() helper as standalone function
- Split server_init.go into server_middlewares.go and server_routes.go
- server_init.go reduced to 76 lines (only initCore remains)

Net: -366 lines, 2 globals eliminated, 20 modules total.
