+++ internal/api/server_middlewares.go
@@ -1,0 +1,151 @@
+package api
+
+import (
+	"strings"
+
+	"github.com/gofiber/fiber/v3"
+	"github.com/gofiber/fiber/v3/middleware/compress"
+	"github.com/gofiber/fiber/v3/middleware/cors"
+	"github.com/gofiber/fiber/v3/middleware/recover"
+	"github.com/gofiber/fiber/v3/middleware/requestid"
+	"github.com/rs/zerolog/log"
+
+	"github.com/nimbleflux/fluxbase/internal/middleware"
+)
+
+func (s *Server) setupMiddlewares() {
+	log.Debug().Msg("Adding requestid middleware")
+	s.app.Use(requestid.New())
+
+	if s.config.Tracing.Enabled && s.tracer != nil && s.tracer.IsEnabled() {
+		log.Debug().Msg("Adding OpenTelemetry tracing middleware")
+		s.app.Use(middleware.TracingMiddleware(middleware.TracingConfig{
+			Enabled:            true,
+			ServiceName:        s.config.Tracing.ServiceName,
+			SkipPaths:          []string{"/health", "/ready", "/metrics"},
+			RecordRequestBody:  false,
+			RecordResponseBody: false,
+		}))
+	}
+
+	if s.config.Metrics.Enabled && s.Metrics.Metrics != nil {
+		log.Debug().Msg("Adding Prometheus metrics middleware")
+		s.app.Use(s.Metrics.Metrics.MetricsMiddleware())
+	}
+
+	log.Debug().Msg("Adding security headers middleware")
+	s.app.Use(func(c fiber.Ctx) error {
+		if strings.HasPrefix(c.Path(), "/admin") {
+			return middleware.AdminUISecurityHeaders()(c)
+		}
+		return middleware.SecurityHeaders()(c)
+	})
+
+	log.Debug().Msg("Adding structured logger middleware")
+	s.app.Use(middleware.StructuredLogger(middleware.StructuredLoggerConfig{
+		SkipPaths:              []string{"/health", "/ready", "/metrics"},
+		SkipSuccessfulRequests: !s.config.Debug,
+	}))
+
+	log.Debug().Msg("Adding recover middleware")
+	s.app.Use(recover.New(recover.Config{
+		EnableStackTrace: s.config.Debug,
+	}))
+
+	corsCredentials := s.config.CORS.AllowCredentials
+	corsOrigins := s.config.CORS.AllowedOrigins
+
+	hasWildcard := false
+	for _, origin := range corsOrigins {
+		if origin == "*" {
+			hasWildcard = true
+			break
+		}
+	}
+
+	if hasWildcard && corsCredentials {
+		log.Warn().Msg("CORS: AllowCredentials disabled because AllowOrigins contains '*' (not allowed per CORS spec)")
+		corsCredentials = false
+	}
+	if !hasWildcard && s.config.PublicBaseURL != "" {
+		found := false
+		for _, origin := range corsOrigins {
+			if origin == s.config.PublicBaseURL {
+				found = true
+				break
+			}
+		}
+		if !found {
+			corsOrigins = append(corsOrigins, s.config.PublicBaseURL)
+			log.Debug().Str("public_url", s.config.PublicBaseURL).Msg("Added public base URL to CORS origins")
+		}
+	}
+	log.Debug().
+		Strs("origins", corsOrigins).
+		Bool("credentials", corsCredentials).
+		Msg("Adding CORS middleware")
+
+	corsConfig := cors.Config{
+		AllowMethods:     s.config.CORS.AllowedMethods,
+		AllowHeaders:     s.config.CORS.AllowedHeaders,
+		ExposeHeaders:    s.config.CORS.ExposedHeaders,
+		AllowCredentials: corsCredentials,
+		MaxAge:           s.config.CORS.MaxAge,
+	}
+
+	if hasWildcard {
+		corsConfig.AllowOriginsFunc = func(origin string) bool {
+			return true
+		}
+	} else {
+		corsConfig.AllowOrigins = corsOrigins
+	}
+
+	s.app.Use(cors.New(corsConfig))
+	log.Debug().Msg("CORS middleware added")
+
+	if len(s.config.Server.AllowedIPRanges) > 0 {
+		log.Info().
+			Int("ranges", len(s.config.Server.AllowedIPRanges)).
+			Strs("ranges", s.config.Server.AllowedIPRanges).
+			Msg("Adding global IP allowlist middleware")
+		s.app.Use(middleware.RequireGlobalIPAllowlist(&s.config.Server))
+	} else {
+		log.Debug().Msg("Global IP allowlist disabled (no ranges configured)")
+	}
+
+	s.app.Use(middleware.DynamicGlobalAPILimiter(s.Auth.Handler.authService.GetSettingsCache(), s.sharedMiddlewareStorage))
+
+	if s.config.Server.BodyLimits.Enabled {
+		bodyLimitConfig := middleware.BodyLimitsFromConfig(
+			s.config.Server.BodyLimits.DefaultLimit,
+			s.config.Server.BodyLimits.RESTLimit,
+			s.config.Server.BodyLimits.AuthLimit,
+			s.config.Server.BodyLimits.StorageLimit,
+			s.config.Server.BodyLimits.BulkLimit,
+			s.config.Server.BodyLimits.AdminLimit,
+			s.config.Server.BodyLimits.MaxJSONDepth,
+		)
+		s.app.Use(middleware.BodyLimitMiddleware(bodyLimitConfig))
+		log.Info().
+			Int64("default", s.config.Server.BodyLimits.DefaultLimit).
+			Int64("rest", s.config.Server.BodyLimits.RESTLimit).
+			Int64("auth", s.config.Server.BodyLimits.AuthLimit).
+			Int64("storage", s.config.Server.BodyLimits.StorageLimit).
+			Int("max_json_depth", s.config.Server.BodyLimits.MaxJSONDepth).
+			Msg("Per-endpoint body limits enabled")
+	}
+
+	idempotencyConfig := middleware.DefaultIdempotencyConfig()
+	idempotencyConfig.DB = s.db.Pool()
+	s.Middleware.Idempotency = middleware.NewIdempotencyMiddleware(idempotencyConfig)
+	s.app.Use(s.Middleware.Idempotency.Middleware())
+	log.Info().
+		Str("header", idempotencyConfig.HeaderName).
+		Dur("ttl", idempotencyConfig.TTL).
+		Msg("Idempotency key support enabled")
+
+	s.app.Use(compress.New(compress.Config{
+		Level: compress.LevelDefault,
+	}))
+}
